// frontend/src/shared/api/generated/models/WalletChallengeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WalletChallengeResponse
 */
export interface WalletChallengeResponse {
    /**
     * 
     * @type {string}
     * @memberof WalletChallengeResponse
     */
    challengeId: string;
    /**
     * 
     * @type {string}
     * @memberof WalletChallengeResponse
     */
    nonce: string;
    /**
     * 
     * @type {Date}
     * @memberof WalletChallengeResponse
     */
    expiresAt: Date;
}

/**
 * Check if a given object implements the WalletChallengeResponse interface.
 */
export function instanceOfWalletChallengeResponse(value: object): value is WalletChallengeResponse {
    if (!('challengeId' in value) || value['challengeId'] === undefined) return false;
    if (!('nonce' in value) || value['nonce'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    return true;
}

export function WalletChallengeResponseFromJSON(json: any): WalletChallengeResponse {
    return WalletChallengeResponseFromJSONTyped(json, false);
}

export function WalletChallengeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): WalletChallengeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'challengeId': json['challengeId'],
        'nonce': json['nonce'],
        'expiresAt': (new Date(json['expiresAt'])),
    };
}

export function WalletChallengeResponseToJSON(json: any): WalletChallengeResponse {
    return WalletChallengeResponseToJSONTyped(json, false);
}

export function WalletChallengeResponseToJSONTyped(value?: WalletChallengeResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'challengeId': value['challengeId'],
        'nonce': value['nonce'],
        'expiresAt': value['expiresAt'].toISOString(),
    };
}