// frontend/src/shared/api/generated/models/WalletLinkRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WalletLinkRequestTonProof } from './WalletLinkRequestTonProof';
import {
    WalletLinkRequestTonProofFromJSON,
    WalletLinkRequestTonProofFromJSONTyped,
    WalletLinkRequestTonProofToJSON,
    WalletLinkRequestTonProofToJSONTyped,
} from './WalletLinkRequestTonProof';
import type { TonNetwork } from './TonNetwork';
import {
    TonNetworkFromJSON,
    TonNetworkFromJSONTyped,
    TonNetworkToJSON,
    TonNetworkToJSONTyped,
} from './TonNetwork';

/**
 * 
 * @export
 * @interface WalletLinkRequest
 */
export interface WalletLinkRequest {
    /**
     * 
     * @type {string}
     * @memberof WalletLinkRequest
     */
    challengeId: string;
    /**
     * 
     * @type {TonNetwork}
     * @memberof WalletLinkRequest
     */
    network: TonNetwork;
    /**
     * 
     * @type {string}
     * @memberof WalletLinkRequest
     */
    walletAddress: string;
    /**
     * 
     * @type {WalletLinkRequestTonProof}
     * @memberof WalletLinkRequest
     */
    tonProof: WalletLinkRequestTonProof;
    /**
     * 
     * @type {string}
     * @memberof WalletLinkRequest
     */
    walletStateInit: string;
    /**
     * 
     * @type {string}
     * @memberof WalletLinkRequest
     */
    publicKey: string;
}



/**
 * Check if a given object implements the WalletLinkRequest interface.
 */
export function instanceOfWalletLinkRequest(value: object): value is WalletLinkRequest {
    if (!('challengeId' in value) || value['challengeId'] === undefined) return false;
    if (!('network' in value) || value['network'] === undefined) return false;
    if (!('walletAddress' in value) || value['walletAddress'] === undefined) return false;
    if (!('tonProof' in value) || value['tonProof'] === undefined) return false;
    if (!('walletStateInit' in value) || value['walletStateInit'] === undefined) return false;
    if (!('publicKey' in value) || value['publicKey'] === undefined) return false;
    return true;
}

export function WalletLinkRequestFromJSON(json: any): WalletLinkRequest {
    return WalletLinkRequestFromJSONTyped(json, false);
}

export function WalletLinkRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): WalletLinkRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'challengeId': json['challengeId'],
        'network': TonNetworkFromJSON(json['network']),
        'walletAddress': json['walletAddress'],
        'tonProof': WalletLinkRequestTonProofFromJSON(json['tonProof']),
        'walletStateInit': json['walletStateInit'],
        'publicKey': json['publicKey'],
    };
}

export function WalletLinkRequestToJSON(json: any): WalletLinkRequest {
    return WalletLinkRequestToJSONTyped(json, false);
}

export function WalletLinkRequestToJSONTyped(value?: WalletLinkRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'challengeId': value['challengeId'],
        'network': TonNetworkToJSON(value['network']),
        'walletAddress': value['walletAddress'],
        'tonProof': WalletLinkRequestTonProofToJSON(value['tonProof']),
        'walletStateInit': value['walletStateInit'],
        'publicKey': value['publicKey'],
    };
}