// frontend/src/shared/api/generated/models/WalletLinkResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WalletState } from './WalletState';
import {
    WalletStateFromJSON,
    WalletStateFromJSONTyped,
    WalletStateToJSON,
    WalletStateToJSONTyped,
} from './WalletState';
import type { TonNetwork } from './TonNetwork';
import {
    TonNetworkFromJSON,
    TonNetworkFromJSONTyped,
    TonNetworkToJSON,
    TonNetworkToJSONTyped,
} from './TonNetwork';

/**
 * 
 * @export
 * @interface WalletLinkResponse
 */
export interface WalletLinkResponse {
    /**
     * 
     * @type {string}
     * @memberof WalletLinkResponse
     */
    walletId: string;
    /**
     * 
     * @type {TonNetwork}
     * @memberof WalletLinkResponse
     */
    network: TonNetwork;
    /**
     * 
     * @type {string}
     * @memberof WalletLinkResponse
     */
    walletAddress: string;
    /**
     * 
     * @type {WalletState}
     * @memberof WalletLinkResponse
     */
    state: WalletState;
    /**
     * 
     * @type {Date}
     * @memberof WalletLinkResponse
     */
    linkedAt: Date;
}



/**
 * Check if a given object implements the WalletLinkResponse interface.
 */
export function instanceOfWalletLinkResponse(value: object): value is WalletLinkResponse {
    if (!('walletId' in value) || value['walletId'] === undefined) return false;
    if (!('network' in value) || value['network'] === undefined) return false;
    if (!('walletAddress' in value) || value['walletAddress'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('linkedAt' in value) || value['linkedAt'] === undefined) return false;
    return true;
}

export function WalletLinkResponseFromJSON(json: any): WalletLinkResponse {
    return WalletLinkResponseFromJSONTyped(json, false);
}

export function WalletLinkResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): WalletLinkResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'walletId': json['walletId'],
        'network': TonNetworkFromJSON(json['network']),
        'walletAddress': json['walletAddress'],
        'state': WalletStateFromJSON(json['state']),
        'linkedAt': (new Date(json['linkedAt'])),
    };
}

export function WalletLinkResponseToJSON(json: any): WalletLinkResponse {
    return WalletLinkResponseToJSONTyped(json, false);
}

export function WalletLinkResponseToJSONTyped(value?: WalletLinkResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'walletId': value['walletId'],
        'network': TonNetworkToJSON(value['network']),
        'walletAddress': value['walletAddress'],
        'state': WalletStateToJSON(value['state']),
        'linkedAt': value['linkedAt'].toISOString(),
    };
}